Split plugin code into module and CLI application
CLI already has a few features, but still needs a lot of clean-up.
